renamed operator package to framework package
