+++ framework/error.go
@@ -1,4 +1,4 @@
-package operator
+package framework
 
 import (
 	"github.com/juju/errgo"
