+++ framework/operator_test.go
@@ -1,4 +1,4 @@
-package operator
+package framework
 
 import (
 	"reflect"
